Move external database call to database activity
