+++ .idea/misc.xml
@@ -3,7 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Harper/OneDrive - James Cook University/2021/Sem 2/CP3407 - Advanced Software Engineering/CP3407-Project/app/src/main/res/layout/activity_apiactivity.xml" value="0.209375" />
+        <entry key="..\:/Users/Harper/OneDrive - James Cook University/2021/Sem 2/CP3407 - Advanced Software Engineering/CP3407-Project/app/src/main/res/layout/activity_db.xml" value="0.209375" />
         <entry key="..\:/Users/Harper/OneDrive - James Cook University/2021/Sem 2/CP3407 - Advanced Software Engineering/CP3407-Project/app/src/main/res/layout/activity_main.xml" value="0.16259057971014493" />
+        <entry key="..\:/Users/Harper/OneDrive - James Cook University/2021/Sem 2/CP3407 - Advanced Software Engineering/CP3407-Project/app/src/main/res/layout/activity_new_word.xml" value="0.209375" />
+        <entry key="..\:/Users/Harper/OneDrive - James Cook University/2021/Sem 2/CP3407 - Advanced Software Engineering/CP3407-Project/app/src/main/res/layout/recyclerview_item.xml" value="0.209375" />
         <entry key="app/src/main/res/layout/activity_apiactivity.xml" value="0.31837606837606836" />
         <entry key="app/src/main/res/layout/activity_db.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.25" />
