+++ app/src/main/res/layout/activity_main.xml
@@ -31,24 +31,5 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <TableLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="ConnectionStatus: "
-            />
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Getting Data..."
-            android:id="@+id/connectionStatus"
-            />
-
-    </TableLayout>
-
 
 </androidx.appcompat.widget.LinearLayoutCompat>
